// src/modules/position/position.enum.ts
export enum PositionStatus {
  ACTIVE = 'Hoạt động',
  INACTIVE = 'Không hoạt động',
}

export enum PositionType {
  DIRECTOR = 'Ban giám đốc',
  MANAGER = 'Quản lý đứng đầu đơn vị',
  STAFF = 'Thành viên Sen Vàng',
  ASSISTANT = 'Trợ lý',
}

export enum PositionSubManager {
  SUB_MANAGER = 'Phó quản lý đơn vị',
  MEMBER = 'Thành viên đơn vị',
}