// src/modules/user/user.enum.ts
export enum UserType {
  ROOT = 'Root',
  ADMIN = 'Admin',
  HR = 'HR',
  CANDIDATE = 'Candidate',
  EMPLOYEE = 'Employee', // ứng viên chấp nhận vào làm việc thì sẽ chuyển từ Candidate sang Employee
  USER = 'User', // ứng viên chấp nhận vào làm việc thì sẽ chuyển từ Candidate sang Employee
  C_B = 'C&B',
}

export enum UserStatus {
  ACTIVE = 'Đang làm việc',
  INACTIVE = 'Không hoạt động',
  DELETED = 'Đã xóa',
}
// ...
export enum UserMovementStatus {
  PENDING = 'Đang chờ',
  APPROVED = 'Đã phê duyệt',
  REJECTED = 'Đã từ chối',
  EFFECTIVE = 'Có hiệu lực',
  CANCELLED = 'Hủy ban hành',
  VIEW = 'Xem',
}